requirements.txt
addict==2.4.0
beautifulsoup4==4.11.1
dotty_dict==1.3.1
imutils==0.5.4
lap==0.4.0
loguru==0.5.3
lvis==0.5.3
matplotlib==3.5.0
mmcls==0.23.2
mmcv==1.6.1
mmcv_full==1.5.3
mmdet==2.25.1
model_archiver==1.0.3
motmetrics==1.2.5
ninja==1.10.2.3
numpy==1.20.3
opencv_python_headless==4.5.4.60
packaging==21.3
pandas==1.3.5
Pillow==9.2.0
pycocotools==2.0.2
pytest==7.1.2
pytorch_sphinx_theme==0.0.19
pywin32==302
PyYAML==6.0
requests==2.26.0
scipy==1.6.2
seaborn==0.11.2
setuptools==58.0.4
tabulate==0.8.9
terminaltables==3.1.10
thop==0.0.31.post2005241907
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.62.3
ts==0.5.1
voice==0.1.0
wandb==0.12.18
win32gui==221.6
xlrd==2.0.1
xlutils==2.0.0
